+++ src/main/java/me/jddev0/ep/block/entity/BatteryBoxBlockEntity.java
@@ -106,7 +106,7 @@ public class BatteryBoxBlockEntity extends BlockEntity implements ExtendedScreen
 
         List<EnergyStorage> consumerItems = new LinkedList<>();
         List<Long> consumerEnergyValues = new LinkedList<>();
-        int consumptionSum = 0;
+        long consumptionSum = 0;
         for(Direction direction:Direction.values()) {
             BlockPos testPos = blockPos.offset(direction);
